requirements.txt
pytest==7.4.3
selenium==4.15.2
webdriver-manager==4.0.1
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-timeout==2.2.0
pytest-json-report==1.5.0
requests==2.31.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
playwright==1.49.0
pytest-playwright==0.5.0
flask==3.0.0
flask-cors==4.0.0

